requirements.txt
pytest==8.3.2
coverage==7.6.1
mutmut==2.5.1
astor==0.8.1
autopep8==2.3.1
openai==1.42.0
langchain-chroma==0.1.3
langchain-core==0.2.37
langchain_openai==0.1.23
langchain_text_splitters==0.2.2
transformers==4.44.2
torch==2.4.0
tqdm==4.66.5
python-dotenv==1.0.1
httpx==0.27.2